+++ common/data_structures/fixed_universe_successor/btrie.h
@@ -73,7 +73,8 @@ class BTrie {
   void Init(size_t u) {
     Clear();
     usize = u;
-    maxh = (u > level_size) ? numeric::ULog2(u - 1) / bits_per_level : 0;
+    maxh =
+        (u > level_size) ? numeric::ULog2(uint64_t(u - 1)) / bits_per_level : 0;
     root.children.resize(maxh > 0 ? level_size : 0, nullptr);
   }
